+++ pkg/network/network_test.go
@@ -14,6 +14,17 @@ import (
 	"github.com/number571/hidden-lake/pkg/response"
 )
 
+func TestError(t *testing.T) {
+	t.Parallel()
+
+	str := "value"
+	err := &SAppError{str}
+	if err.Error() != errPrefix+str {
+		t.Error("incorrect err.Error()")
+		return
+	}
+}
+
 func TestPanicNode(t *testing.T) {
 	t.Parallel()